refactor(editor): Moved persistense directory into infrastructure
