+++ module/editor/src/Infrastructure/Persistence/Projector/DbalProductDraftAppliedEventProjector.php
@@ -7,7 +7,7 @@
 
 declare(strict_types = 1);
 
-namespace Ergonode\Editor\Persistence\Projector;
+namespace Ergonode\Editor\Infrastructure\Persistence\Projector;
 
 use Doctrine\DBAL\Connection;
 use Doctrine\DBAL\DBALException;
@@ -15,7 +15,7 @@ use Ergonode\Editor\Domain\Event\ProductDraftApplied;
 
 /**
  */
-class ProductDraftAppliedEventProjector
+class DbalProductDraftAppliedEventProjector
 {
     private const PRODUCT_TABLE = 'product';
     private const DRAFT_TABLE = 'designer.draft';
